+++ src/domain/command.model.ts
