+++ src/domain/target-number.model.ts
