+++ test/target-number.model.test.ts
@@ -1,5 +1,5 @@
 import { MAX_NUMBER, MIN_NUMBER } from '../src/const/num-range';
-import { TargetNumber } from '../src/domain/TargetNumber';
+import { TargetNumber } from '../src/domain/target-number.model';
 
 describe('TargetNumber 클래스 테스트', () => {
     it('세자리 숫자로 이루어진 문자열로 새 인스턴스를 생성하면 해당 숫자로 이루어진 number 배열이 생성된다', () => {
